load_model & Deployment/requirements.txt
fastapi
pillow
numpy
tensorflow
uvicorn
gunicorn
python-multipart